pipeline/requirements.txt
open-clip-torch==2.24.0
torch>=2.9.0
torchvision>=0.20.0
Pillow>=11.0.0
qdrant-client==1.9.1
imagehash==4.3.1
datasets==2.19.1
cloudinary==1.40.0
httpx==0.27.0
tqdm==4.66.4
python-dotenv==1.0.1
pandas==2.2.2
